feat: handle try to add same file to more than once that produce the extension to crash when parsing JSON context file

+++ src/comands.ts
@@ -2,6 +2,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 import * as vscode from 'vscode';
 import { SimpleDataProvider } from './dataProvider';
+import { SingletonOutputChannel } from './loggerChannel';
 
 export async function showSectionPicker(): Promise<string | undefined> {
 	const workspaceRoot = vscode.workspace.rootPath;
@@ -9,7 +10,17 @@ export async function showSectionPicker(): Promise<string | undefined> {
 
 	if (fs.existsSync(configPath)) {
 		const config = JSON.parse(fs.readFileSync(configPath, 'utf-8'));
+		 
 		const sections = Object.keys(config.contexts);
+
+		if (sections.length===0){
+			let message='File contexts.json in .vscode does not contain any folder. Create firsts a folder to add the selected file.';
+			SingletonOutputChannel.appendLine(message);
+
+			vscode.window.showWarningMessage(message);
+			return undefined;
+		}
+
 		return await vscode.window.showQuickPick(sections, {
 			placeHolder: 'Select a context to add the file to'
 		});
@@ -29,6 +40,10 @@ export function addFileToConfig(filePath: string, section: string, dataProvider:
 		const fileContent = JSON.parse(fs.readFileSync(configPath, 'utf-8'));
 
 		if (fileContent.contexts[section]) {
+			if (fileContent.contexts[section].includes(relativePath)){
+				vscode.window.showWarningMessage(`File:${relativePath} already exists in context ${section}.`);
+				return;
+			}
 			fileContent.contexts[section].push(relativePath);
 			fs.writeFileSync(configPath, JSON.stringify(fileContent, null, 4));
 			dataProvider.refresh();
